// frontend/src/types/themeTypes/ThemeType.ts
export type ThemeTypes = {
  name: string;
  dark: boolean;
  variables?: object;
  colors: {
    primary?: string;
    secondary?: string;
    primary_lighten_5?: string;
    primary_lighten_4?: string;
    primary_lighten_3?: string;
    primary_lighten_2?: string;
    primary_lighten_1?: string;
    primary_darken_1?: string;
    primary_darken_2?: string;
    primary_darken_3?: string;
    primary_darken_4?: string;
    secondary_darken_5?: string;
    secondary_darken_4?: string;
    secondary_darken_3?: string;
    secondary_darken_2?: string;
    secondary_darken_1?: string;
    secondary_lighten_1?: string;
    secondary_lighten_2?: string;
    secondary_lighten_3?: string;
    secondary_lighten_4?: string;
    secondary_lighten_5?: string;
    info?: string;
    success?: string;
    accent?: string;
    warning?: string;
    error?: string;
    simplePDCA?: string;
    lightprimary?: string;
    lightsecondary?: string;
    lightsuccess?: string;
    lighterror?: string;
    lightwarning?: string;
    darkprimary?: string;
    darksecondary?: string;
    darkText?: string;
    lightText?: string;
    borderLight?: string;
    inputBorder?: string;
    containerBg?: string;
    surface?: string;
    background?: string;
    'on-surface-variant'?: string;
    facebook?: string;
    twitter?: string;
    linkedin?: string;
    gray100?: string;
    primary200?: string;
    secondary200?: string;
    gray50?: string;
    gray200?: string;
    gray300?: string;
    gray400?: string;
    gray500?: string;
    gray600?: string;
    gray700?: string;
    gray800?: string;
    gray900?: string;
  };
};